pyproject.toml
[project]
authors = [
  {name = "andyzhou", email = "andyforever0108@outlook.com"},
]
dependencies = [
  "flask~=2.0",
  "apiflask~=0.10",
  "flask-sqlalchemy~=2.5",
  "flask-wtf~=0.15",
  "wtforms-sqlalchemy~=0.2",
  "flask-login~=0.5",
  "bootstrap-flask~=1.8",
  "flask-compress~=1.10"]
description = ""
dynamic = ["classifiers"]
license = {text = "MIT"}
name = "djask"
requires-python = ">=3.7"
version = "0.1.0"

[project.urls]
homepage = ""

[project.optional-dependencies]
[build-system]
build-backend = "pdm.pep517.api"
requires = ["pdm-pep517"]

[tool]
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
  "black<22,>=21.9b0",
  "pytest~=6.2",
  "pytest-cov~=3.0",
]

[tool.pytest.ini_options]
addopts = "--cov=djask"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
omit = ["*.html"]
source = ["djask"]
